Refactor music player webview: - Updated HTML structure to enhance album art display and track information. - Removed old JavaScript and CSS files, replacing them with modularized versions for better maintainability. - Introduced new CSS styles for improved UI aesthetics and responsiveness. - Implemented a new JavaScript architecture with separate utility functions for music logic, UI updates, and time management. - Enhanced progress bar functionality and added event listeners for control buttons. - Added support for dynamic updates of track information and playback status.
