--- src/common/ui/webview/musicPlayer_old.css
@@ -1,287 +1,0 @@
-/* VS Code Music Player - Compact UI Styles */
-
-body {
-    font-family: var(--vscode-font-family);
-    color: var(--vscode-foreground);
-    background-color: var(--vscode-editor-background);
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-    font-size: 13px;
-}
-
-/* Utility classes */
-.hidden {
-    display: none !important;
-}
-
-.music-container {
-    max-width: 400px;
-    margin: 0 auto;
-}
-
-/* No Music State */
-.no-music-container {
-    text-align: center;
-    padding: 32px 16px;
-    opacity: 0.7;
-}
-
-.music-icon {
-    font-size: 48px;
-    margin-bottom: 16px;
-}
-
-.no-music-container h2 {
-    margin: 0 0 8px 0;
-    font-size: 18px;
-    font-weight: 600;
-    color: var(--vscode-foreground);
-}
-
-.no-music-container p {
-    margin: 0 0 16px 0;
-    color: var(--vscode-descriptionForeground);
-    font-size: 13px;
-}
-
-.supported-players {
-    display: flex;
-    gap: 8px;
-    justify-content: center;
-    flex-wrap: wrap;
-}
-
-.supported-players span {
-    background: var(--vscode-badge-background);
-    color: var(--vscode-badge-foreground);
-    padding: 4px 8px;
-    border-radius: 4px;
-    font-size: 11px;
-    font-weight: 500;
-}
-
-/* Music Info Display */
-.music-info {
-    background: var(--vscode-sideBar-background);
-    border: 1px solid var(--vscode-sideBar-border);
-    border-radius: 8px;
-    padding: 16px;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-}
-
-.music-info.hidden {
-    display: none;
-}
-
-.track-container {
-    display: flex;
-    gap: 16px;
-    margin-bottom: 16px;
-    align-items: flex-start;
-}
-
-.artwork-container {
-    width: 80px;
-    height: 80px;
-    flex-shrink: 0;
-    border-radius: 6px;
-    overflow: hidden;
-    background: var(--vscode-input-background);
-    border: 1px solid var(--vscode-input-border);
-    position: relative;
-}
-
-.artwork {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 6px;
-}
-
-.artwork-placeholder {
-    width: 100%;
-    height: 100%;
-    background: var(--vscode-input-background);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 24px;
-    color: var(--vscode-input-placeholderForeground);
-    border: 2px solid var(--vscode-input-border);
-    border-radius: 6px;
-}
-
-.track-details {
-    flex: 1;
-    text-align: left;
-    min-width: 0;
-    overflow: hidden;
-}
-
-.track-title {
-    font-weight: 600;
-    font-size: 16px;
-    margin-bottom: 8px;
-    color: var(--vscode-foreground);
-    line-height: 1.3;
-    overflow: hidden;
-    white-space: nowrap;
-    animation: scroll-text 15s linear infinite;
-    padding-right: 20px;
-}
-
-.track-artist {
-    font-size: 14px;
-    color: var(--vscode-descriptionForeground);
-    margin-bottom: 4px;
-    font-weight: 500;
-    overflow: hidden;
-    white-space: nowrap;
-    animation: scroll-text 12s linear infinite;
-    padding-right: 20px;
-}
-
-.track-album {
-    font-size: 12px;
-    color: var(--vscode-disabledForeground);
-    font-style: italic;
-    opacity: 0.8;
-}
-
-@keyframes scroll-text {
-    0% {
-        transform: translateX(0%);
-    }
-
-    20% {
-        transform: translateX(0%);
-    }
-
-    80% {
-        transform: translateX(calc(-100% + 200px));
-    }
-
-    100% {
-        transform: translateX(calc(-100% + 200px));
-    }
-}
-
-.track-title:hover,
-.track-artist:hover {
-    animation-play-state: paused;
-}
-
-.status-indicator {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 20px;
-    height: 20px;
-    border-radius: 50%;
-    font-size: 10px;
-    margin-bottom: 8px;
-    font-weight: bold;
-}
-
-.status-playing {
-    background: var(--vscode-testing-iconPassed);
-    color: white;
-}
-
-.status-paused {
-    background: var(--vscode-testing-iconQueued);
-    color: white;
-}
-
-.status-stopped {
-    background: var(--vscode-testing-iconFailed);
-    color: white;
-}
-
-/* Progress Bar */
-.progress-container {
-    margin: 16px 0;
-}
-
-.progress-bar {
-    width: 100%;
-    height: 4px;
-    background: var(--vscode-progressBar-background);
-    border-radius: 2px;
-    overflow: hidden;
-    margin-bottom: 8px;
-}
-
-.progress-fill {
-    height: 100%;
-    background: var(--vscode-progressBar-foreground);
-    border-radius: 2px;
-    transition: width 0.5s ease;
-    width: 0%;
-}
-
-.progress-time {
-    display: flex;
-    justify-content: space-between;
-    font-size: 11px;
-    color: var(--vscode-descriptionForeground);
-    font-weight: 500;
-}
-
-/* Controls */
-.controls {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 12px;
-    margin-top: 16px;
-}
-
-.control-btn {
-    width: 36px;
-    height: 36px;
-    border: none;
-    border-radius: 6px;
-    background: var(--vscode-button-background);
-    color: var(--vscode-button-foreground);
-    cursor: pointer;
-    font-size: 14px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: all 0.2s ease;
-}
-
-.control-btn:hover {
-    background: var(--vscode-button-hoverBackground);
-    transform: translateY(-1px);
-}
-
-.control-btn:active {
-    transform: translateY(0);
-}
-
-.play-pause-btn {
-    width: 44px;
-    height: 44px;
-    font-size: 16px;
-    background: var(--vscode-button-background);
-}
-
-.play-pause-btn:hover {
-    background: var(--vscode-button-hoverBackground);
-}
-
-/* Responsive adjustments */
-@media (max-width: 300px) {
-    .track-container {
-        flex-direction: column;
-        align-items: center;
-        text-align: center;
-    }
-
-    .track-details {
-        text-align: center;
-    }
-}
