+++ src/common/ui/webview/static/css/musicPlayer.css
