+++ src/common/ui/webview/static/js/utils/showMusic.js
@@ -1,0 +1,28 @@
+
+
+function updateStatusIndicator(status) {
+    const statusIndicator = document.getElementById('status-indicator');
+    if (!statusIndicator) {
+        return;
+    }
+
+    // Remove existing status classes
+    statusIndicator.className = 'status-indicator';
+
+    // Add appropriate status class and icon
+    if (status === 'playing') {
+        statusIndicator.classList.add('status-playing');
+        statusIndicator.textContent = '▶';
+    } else if (status === 'paused') {
+        statusIndicator.classList.add('status-paused');
+        statusIndicator.textContent = '⏸';
+    } else {
+        statusIndicator.classList.add('status-stopped');
+        statusIndicator.textContent = '⏹';
+    }
+}
+
+
+export {
+    updateStatusIndicator,
+};
